+++ packages/types/src/countries.ts
@@ -1,0 +1,5 @@
+import { countries } from 'country-flag-icons';
+
+export const COUNTRY_CODES = countries;
+
+export default COUNTRY_CODES;